backend/pyproject.toml
[project]
name = "sumi-backend"
version = "0.1.0"
description = "Sumi: Artistic Infographic Generator - Backend Engine"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=2.1.0",
    "pydantic-settings>=2.6.0",
    "google-genai>=1.0.0",
    "fal-client>=0.5.0",
    "anthropic>=0.40.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
    "Pillow>=11.0.0",
    "python-multipart>=0.0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sumi"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
